+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2026-04-30
+
+- More robust temporary token handling
+- Better handling for asynchronous API requests
+
 ## [1.0.0] - 2025-10-11
 
 Initial release